ml_service/requirements.txt
fastapi
uvicorn
tensorflow==2.13.0
keras==2.13.1
numpy<2
pillow
python-multipart
python-jose
requests
